+++ phantom/cython/dijkstra_priority_queue.hpp
@@ -2,6 +2,6 @@
 #include <queue>
 #include <utility>
 
-using coords = std::pair<int,int>;
+using coords = std::pair<Py_ssize_t,Py_ssize_t>;
 using item = std::pair<float,coords>;
 using cpp_pq = std::priority_queue<item,std::vector<item>,std::function<bool(item,item)>>;
